+++ hack/tools/confdocs/confdocs.go
@@ -20,16 +20,16 @@ import (
 	"path/filepath"
 	"regexp"
 	"sort"
-	"strconv"
 	"strings"
 	"text/template"
 
-	"github.com/go-andiamo/splitter"
 	"github.com/mattn/go-isatty"
 	"go.elastic.co/ecszap"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"golang.org/x/tools/go/packages"
+
+	"github.com/cerbos/cerbos/hack/tools/confdocs/structtag"
 )
 
 const (
@@ -399,7 +399,7 @@ func parseTag(tag string) (*TagInfo, error) {
 		return nil, errTagNotExists
 	}
 
-	t, err := Parse(tag[1 : len(tag)-1])
+	t, err := structtag.Parse(tag[1 : len(tag)-1])
 	if err != nil {
 		return nil, fmt.Errorf("structtag failed to parse tags: %w", err)
 	}
@@ -542,123 +542,4 @@ func doInitLogging(level string) {
 	zap.RedirectStdLog(l.Named("stdlog"))
 
 	logger = l.Sugar()
-}
-
-var (
-	errTagSyntax      = errors.New("bad syntax for struct tag pair")
-	errTagKeySyntax   = errors.New("bad syntax for struct tag key")
-	errTagValueSyntax = errors.New("bad syntax for struct tag value")
-	errTagNotExist    = errors.New("tag does not exist")
-)
-
-type Tags struct {
-	tags []*Tag
-}
-
-type Tag struct {
-	// Key is the tag key, such as json, xml, etc..
-	// i.e: `json:"foo,omitempty". Here key is: "json"
-	Key string
-
-	// Name is a part of the value
-	// i.e: `json:"foo,omitempty". Here name is: "foo"
-	Name string
-
-	// Options is a part of the value. It contains a slice of tag options i.e:
-	// `json:"foo,omitempty". Here options is: ["omitempty"]
-	Options []string
-}
-
-// Parse and others are from https://github.com/fatih/structtag/blob/2977b8db49bbf5ae2e0ae2be55e43d2c1798fc03/tags.go#L42-L125
-func Parse(tag string) (*Tags, error) {
-	var tags []*Tag
-
-	hasTag := tag != ""
-	for tag != "" {
-		i := 0
-		for i < len(tag) && tag[i] == ' ' {
-			i++
-		}
-		tag = tag[i:]
-		if tag == "" {
-			break
-		}
-
-		i = 0
-		for i < len(tag) && tag[i] > ' ' && tag[i] != ':' && tag[i] != '"' && tag[i] != 0x7f {
-			i++
-		}
-
-		if i == 0 {
-			return nil, errTagKeySyntax
-		}
-		if i+1 >= len(tag) || tag[i] != ':' {
-			return nil, errTagSyntax
-		}
-		if tag[i+1] != '"' {
-			return nil, errTagValueSyntax
-		}
-
-		key := tag[:i]
-		tag = tag[i+1:]
-
-		i = 1
-		for i < len(tag) && tag[i] != '"' {
-			if tag[i] == '\\' {
-				i++
-			}
-			i++
-		}
-		if i >= len(tag) {
-			return nil, errTagValueSyntax
-		}
-
-		qvalue := tag[:i+1]
-		tag = tag[i+1:]
-
-		value, err := strconv.Unquote(qvalue)
-		if err != nil {
-			return nil, errTagValueSyntax
-		}
-
-		split, err := splitter.NewSplitter(',', splitter.SquareBrackets)
-		if err != nil {
-			return nil, fmt.Errorf("failed to create a new splitter: %w", err)
-		}
-
-		res, err := split.Split(value)
-		if err != nil {
-			return nil, fmt.Errorf("failed to split value %s: %w", value, err)
-		}
-
-		name := res[0]
-		options := res[1:]
-		if len(options) == 0 {
-			options = nil
-		}
-
-		tags = append(tags, &Tag{
-			Key:     key,
-			Name:    name,
-			Options: options,
-		})
-	}
-
-	if hasTag && len(tags) == 0 {
-		return nil, nil
-	}
-
-	return &Tags{
-		tags: tags,
-	}, nil
-}
-
-func (t *Tags) Get(key string) (*Tag, error) {
-	for _, tag := range t.tags {
-		if tag.Key == key {
-			return tag, nil
-		}
-	}
-
-	return nil, errTagNotExist
 }
